requirements.txt
Django>=5.0.0,<5.0.3
djangorestframework>=3.14.0,<3.15.0
psycopg2>=2.8.6,<2.9.9
drf-spectacular>=0.20.0,<0.27.1